코인 업데이트 시 이벤트를 발생시켜 들어와있는 매수 매도 주문을 체결한다.

+++ backend/src/main/java/com/kmbbj/backend/feature/exchange/service/TransactionServiceImpl.java
@@ -66,7 +66,7 @@ public class TransactionServiceImpl implements TransactionService {
     }
 
     @Override
-    public void matchOrders(Long coinId, BigDecimal price) {
+    public void matchOrders(Long coinId, Long price) {
         executionAllMatchingOrder.matchOrders(coinId, price);
     }
